:sparkles: this adds the events section to the site
I used a hook (I got the 'inspiration' from the gitpod website - thank you @mikenikles for doing the big part of this hook 🙃) to provide the list of the events with their frontmatter data to the index events page. I also fixed the name of the new layout - it got broken with the mdsvex installation, but I fixed it.